[FIX] Issue#25 (color sensor don't work properly) [UPDATE] Issue#21 (upTime in status message are now in mS)

+++ algobot_main.h
@@ -112,7 +112,7 @@ typedef struct tsensor{
 
 typedef struct tSystem{
 	char name[32];
-        int startUpTime;
+        long startUpTime;
         char firmwareVersion[32];
         char mcuVersion[32];
         char HWrevision[32];
